+++ eq_report_pattern/__manifest__.py
@@ -22,7 +22,7 @@
 {
     'name': 'Equitania Dokumenten Bausteine',
     'license': 'AGPL-3',
-    'version': '1.0.28',
+    'version': '1.0.29',
     'description': """
         Improves the odoo document templating, overwrites head- and foot-texttemplates
     """,