+++ eq_report_pattern/models/document_template.py
@@ -28,6 +28,7 @@ class eq_document_template(models.Model):
     _name = "eq.document.template"
     
     name = fields.Char(string="Name", translate=True, required=True)
+    eq_model = fields.Selection(selection=[('sale.order', 'Sale Order'), ('purchase.order', 'Purchase Order')],string="Model")
     eq_quote_line = fields.One2many('eq.sale.quote.line', 'quote_id', 'Quote Template Lines', copy=True)
     eq_header = fields.Html(string="Header", translate=True)
     eq_footer = fields.Html(string="Footer", translate=True)