+++ eq_report_pattern/models/eq_sale_quote_line.py
@@ -18,13 +18,7 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
-
-
-
 from openerp import models, fields, api, _
-import datetime
-import time
-from openerp.tools import DEFAULT_SERVER_DATETIME_FORMAT
 
 class eq_sale_quote_line(models.Model):
     _name = "eq.sale.quote.line"
@@ -54,3 +48,22 @@ class eq_sale_quote_line(models.Model):
                 values = self._inject_quote_description(values)
                 res = super(eq_sale_quote_line, self).create(values)
                 return res
+
+    @api.onchange('product_id')
+    def _onchange_product_id(self):
+        self.ensure_one()
+        if self.product_id:
+            if self.quote_id.eq_model == 'purchase.order':
+                self.price_unit = self.product_id.standard_price
+            else:
+                self.price_unit = self.product_id.lst_price
+            name = self.product_id.name_get()[0][1]
+            if self.product_id.description_sale:
+                name += '\n' + self.product_id.description_sale
+            self.name = name
+            self.product_uom_id = self.product_id.uom_id.id
+
+    @api.onchange('product_uom_id')
+    def _onchange_product_uom(self):
+        if self.product_id and self.product_uom_id and self.quote_id.eq_model != 'purchase.order':
+            self.price_unit = self.product_id.uom_id._compute_price(self.product_id.lst_price, self.product_uom_id)